requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.47
typing_extensions==4.15.0
urllib3==2.6.3
